Use uint64_t for compatibility

+++ lib/fsttable/fsttable.h
@@ -66,7 +66,7 @@ public:
 	    return string_encoding;
 	}
 
-	void BufferToVec(unsigned long long nrOfElements, unsigned long long startElem, unsigned long long endElem, unsigned long long vecOffset,
+	void BufferToVec(uint64_t nrOfElements, uint64_t startElem, uint64_t endElem, uint64_t vecOffset,
     unsigned int* sizeMeta, char* buf);
 
 	const char* GetElement(unsigned long long elementNr);
